@@ -1,7 +1,8 @@
 {
-  "name": "@example/starter",
-  "version": "0.0.1",
   "private": true,
+  "name": "@example/starter",
+  "type": "module",
+  "version": "0.0.1",
   "scripts": {
     "dev": "astro dev",
     "start": "astro dev",
@@ -12,39 +13,39 @@
   },
   "devDependencies": {
     "@astrojs/renderer-vue": "^0.4.0",
-    "@astrojs/tailwind": "^2.0.1",
-    "@astrojs/vue": "^1.0.0",
-    "@cooklang/cooklang-ts": "^1.2.0",
-    "@types/node": "^18.7.18",
-    "@typescript-eslint/eslint-plugin": "^5.38.0",
-    "@typescript-eslint/parser": "^5.38.0",
-    "astro": "^1.2.7",
-    "eslint": "^8.23.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-astro": "^0.19.0",
+    "@astrojs/tailwind": "^3.0.1",
+    "@astrojs/vue": "^2.0.1",
+    "@cooklang/cooklang-ts": "^1.2.3",
+    "@types/node": "^18.13.0",
+    "@typescript-eslint/eslint-plugin": "^5.51.0",
+    "@typescript-eslint/parser": "^5.51.0",
+    "astro": "^2.0.7",
+    "eslint": "^8.33.0",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-plugin-astro": "^0.23.0",
     "eslint-plugin-prettier": "4.2.1",
-    "eslint-plugin-vue": "^9.5.1",
+    "eslint-plugin-vue": "^9.9.0",
     "path-browserify": "^1.0.1",
-    "postcss": "^8.4.16",
-    "prettier": "^2.7.1",
-    "prettier-plugin-astro": "^0.5.3",
-    "tailwindcss": "^3.1.8",
-    "typescript": "^4.8.2",
+    "postcss": "^8.4.21",
+    "prettier": "^2.8.3",
+    "prettier-plugin-astro": "^0.8.0",
+    "tailwindcss": "^3.2.4",
+    "typescript": "^4.9.5",
     "vite-plugin-eslint": "^1.8.1"
   },
   "dependencies": {
-    "@esbuild-plugins/node-globals-polyfill": "^0.1.1",
-    "@esbuild-plugins/node-modules-polyfill": "^0.1.4",
-    "@fontsource/league-spartan": "^4.5.11",
-    "@rollup/plugin-inject": "^4.0.4",
-    "@tailwindcss/typography": "^0.5.7",
-    "autoprefixer": "^10.4.11",
-    "daisyui": "^2.29.0",
+    "@esbuild-plugins/node-globals-polyfill": "^0.2.3",
+    "@esbuild-plugins/node-modules-polyfill": "^0.2.2",
+    "@fontsource/league-spartan": "^4.5.13",
+    "@rollup/plugin-inject": "^5.0.3",
+    "@tailwindcss/typography": "^0.5.9",
+    "autoprefixer": "^10.4.13",
+    "daisyui": "^2.50.0",
     "node-stdlib-browser": "^1.2.0",
-    "pinia": "^2.0.21",
+    "pinia": "^2.0.30",
     "resolve": "^1.22.1",
     "rollup-plugin-browserify-transform": "^1.0.1",
     "ts-node": "^10.9.1",
-    "vue": "^3.2.37"
+    "vue": "^3.2.47"
   }
 }
