@@ -1,6 +1,9 @@
-const defaultTheme = require('tailwindcss/defaultTheme')
+import defaultTheme from 'tailwindcss/defaultTheme.js'
 
-module.exports = {
+import typo from '@tailwindcss/typography'
+import daisyui from 'daisyui'
+
+export default {
   content: ['./src/**/*.{astro,html,js,jsx,md,mdx,svelte,ts,tsx,vue}'],
   theme: {
     extend: {
@@ -9,5 +12,5 @@ module.exports = {
       },
     },
   },
-  plugins: [require('@tailwindcss/typography'), require('daisyui')],
+  plugins: [typo, daisyui],
 }
