@@ -1,20 +1,17 @@
 <template>
-  <div
-    class="recipe-preview card card-bordered border-dashed bg-base-100 shadow-sm"
-  >
-    <div class="card-body prose">
+    <span>
       <h3 class="title font-display text-2xl font-light">
         {{ recipe.metadata?.title }}
       </h3>
-      <p class="servings">Servings: {{ recipe.metadata?.servings ?? '??' }}</p>
+
+      <!-- <p class="servings">Servings: {{ recipe.metadata?.servings ?? '??' }}</p>
 
       <div class="card-actions justify-end">
         <a :href="recipeUrl" class="btn btn-outline btn-primary btn-sm"
           >Read more</a
         >
-      </div>
-    </div>
-  </div>
+      </div> -->
+    </span>
 </template>
 
 <script lang="ts">
